Fix bug where timeouts ignore queue time
Time spent waiting to be written was not being counted against
the timeout. We now track the time spent in queue and deduct it
from the timeout once the message is written.

We opted to keep the timeout timer starting on write in order to
generate the opaque as late as possible -- this will allow
for queued commands to be transferred to a new connection once
reconnect is implemented.